requirements.test.txt
# Requirements for running qa and tests

# flake8 - syntax checking
mccabe==0.2.1
pep8==1.5.7
pyflakes==0.8.1
flake8==2.2.2

# coverage - line and branch coverage
coverage==3.7.1

# nose - improved test runner
nose==1.3.3
django-nose==1.2
